requirements.txt
python-telegram-bot==13.7
urllib3<2.0.0
APScheduler>=3.10.1
python-dotenv==1.0.0
supabase==2.3.0
openai==1.12.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
pytz==2024.1
six>=1.10.0
langchain==0.1.0
langchain-community==0.0.13
pandas==2.2.0
numpy==1.26.3
httpx>=0.24.0,<0.25.0 